requirements.txt
# Core Dependencies
#python>=3.8,<3.10
numpy>=1.26.0
pandas>=2.1.1
scikit-learn>=1.3.0
scipy>=1.11.0

# Data Fetching
yfinance>=0.2.31
alpaca-trade-api>=3.0.0
polygon-api-client>=1.13.3
aiohttp>=3.8.5
requests>=2.31.0

# Machine Learning
tensorflow>=2.15.0
prophet>=1.1.5
alibi-detect>=0.11.1
changefinder==0.3.1
statsmodels>=0.14.0

# Technical Analysis
ta-lib>=0.4.24
vectorbt>=0.25.4

# Sentiment Analysis
textblob>=0.17.1
praw>=7.7.1
twscrape>=0.5.0
newsapi-python>=0.2.7
transformers>=4.35.2

# Visualization
plotly>=5.18.0
streamlit>=1.28.0
matplotlib>=3.7.0

# AWS Integration (Optional)
#boto3>=1.28.62
#awscli>=1.29.0

# Database
sqlite3>=2.6.0
psycopg2>=2.9.0

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.6
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

#tensor flow and tensorflow probability

alibi-detect[tensorflow]
tensorflow_probability